simplifying BattleManager into BattleController and the InitiativeController into TurnController
