+++ src/controllers/battle/RoundState.cs
@@ -1,10 +1,7 @@
 namespace DiceRolling.Controllers;
 
-/// Define os possíveis estados da batalha.
-public enum BattleState {
-    Start,
-    InProgress,
-    End,
+/// Define os possíveis estados dos rounds.
+public enum RoundState {
     RoundStart,
     ActionsDeclaration,
     TurnsResolution,
